backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.12
idna==3.10
pydantic==2.11.7
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn[standard]
requests
sqlmodel
psycopg[binary]
sqlalchemy
langchain
langchain-openai
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
certifi
Pillow
huggingface-hub
supabase
aiofiles
uuid
opencv-python-headless
numpy
